CalibrateAnalogControls.cs : Added example to calibrate left trigger in function sendCalibrationTable(). This is Alpha code and can be used as a reference but the entire feature is not yet complete. Changed the calibration table from type int[] to byte[].
XBoxController_v100.cs : Added XBGamePad_t to the driver. This will eventually depricate isWired. New XBoxControllerConfiguration class will handle all calibration and configuration functions for the driver. One of its main collaborators is XBoxCalibration. This class manages the configuration page in EEPROM as well as the calibration records for each of the analog controls.